Moved logic into own projects (2nd try)
