+++ CommonApp/AboutBoxMain.eo.resx
