+++ CommonApp/AboutBoxMain.resx
