+++ CommonApp/ProjectProperties.cs
@@ -1,4 +1,4 @@
-﻿namespace WindowStreamer.Common;
+﻿namespace CommonApp;
 
 public static class ProjectProperties
 {
