+++ CommonApp/ToolStripLabelSink.cs
