+++ ServerApp/ServerApp.csproj
@@ -51,7 +51,7 @@
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
-    <PackageReference Include="Serilog" Version="2.12.0" />
+    <PackageReference Include="Serilog" Version="3.0.1" />
     <PackageReference Include="Serilog.Sinks.Console" Version="4.1.0" />
     <PackageReference Include="Serilog.Sinks.File" Version="5.0.0" />
     <PackageReference Include="StyleCop.Analyzers" Version="1.1.118">
@@ -73,7 +73,8 @@
     <Content Include="Resources\ApplicationIcon.ico" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\WindowStreamer.Common\WindowStreamer.Common.csproj" />
+    <ProjectReference Include="..\CommonApp\CommonApp.csproj" />
     <ProjectReference Include="..\WindowStreamer.Protocol\WindowStreamer.Protocol.csproj" />
+    <ProjectReference Include="..\WindowStreamer.Server\WindowStreamer.Server.csproj" />
   </ItemGroup>
 </Project>
