+++ WindowStreamer.Server/WindowServer.cs
@@ -1,7 +1,6 @@
 ﻿using Serilog;
 using System.Diagnostics;
 using System.Drawing;
-using System.Drawing.Imaging;
 using System.Net;
 using System.Net.Sockets;
 using System.Runtime.InteropServices;
